requirements.txt
adal==1.2.6
astroid==2.4.1
azure-common==1.1.26
azure-core==1.12.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==12.1.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.8.0
azureml-core==1.24.0.post2
azureml-dataprep==2.11.2
azureml-dataprep-native==30.0.0
azureml-dataprep-rslex==1.9.1
azureml-dataset-runtime==1.25.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
celluloid==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
commonmark==0.9.1
contextlib2==0.6.0.post1
cryptography==3.4.6
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
distro==1.5.0
docker==4.4.4
dotnetcore2==2.1.20
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.3
jmespath==0.10.0
JSAnimation==0.1
jsonpickle==2.0.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
msal==1.10.0
msal-extensions==0.2.2
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.19.3
oauthlib==3.1.0
pandas==1.2.3
parso==0.8.1
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
portalocker==1.7.1
prompt-toolkit==3.0.17
pyarrow==1.0.1
pyasn1==0.4.8
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pylint==2.5.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==227
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rich==9.13.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scipy==1.6.2
seaborn==0.11.1
SecretStorage==3.3.1
six==1.15.0
toml==0.10.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
websocket-client==0.58.0
wrapt==1.12.1
zope.interface==5.2.0
